+++ app/src/main/java/edu/uw/npctracker/core/Character.kt
@@ -1,4 +1,4 @@
-package edu.uw.npctracker
+package edu.uw.npctracker.core
 
 import java.time.Instant
