+++ app/src/main/java/edu/uw/npctracker/core/Tag.kt
@@ -1,4 +1,4 @@
-package edu.uw.npctracker
+package edu.uw.npctracker.core
 
 /*
  * @author lstrobel
